requirements.txt
pydantic-ai[groq]>=0.0.14
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
python-dotenv>=1.0.0
